- Delete duplicate activities: LoginActivity and SignUpActivity - Order all the dimens, strings and id's in both of the activities - make contacts between the activities: login, sign up and home page.
