+++ app/src/main/java/com/example/secondproject/LoginActivity.java
@@ -8,7 +8,6 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.text.TextUtils;
 import android.view.View;
-import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.Toast;
@@ -96,15 +95,6 @@ public class LoginActivity extends AppCompatActivity {
         });
 
     }
-
-    public void launchSignUpActivity(View view) {
-        Intent intent = new Intent(this, SignUpActivity.class);
-        startActivity(intent);
-    }
-
-    public void launchHomePageActivity(View view) {
-        Intent intent = new Intent(this, HomePage.class);
-        startActivity(intent);
 
     private void SignIn() {
         Intent signInIntent = mGoogleSignInClient.getSignInIntent();
@@ -205,4 +195,15 @@ public class LoginActivity extends AppCompatActivity {
         });
 
     }
+
+    public void launchSignUpActivity(View view) {
+        Intent intent = new Intent(this, SignUpActivity.class);
+        startActivity(intent);
+    }
+
+    public void launchHomePageActivity(View view) {
+        Intent intent = new Intent(this, HomePageActivity.class);
+        startActivity(intent);
+    }
+
 }
