+++ app/src/main/res/layout/activity_friend_shazamzams.xml
@@ -40,7 +40,7 @@
         app:layout_constraintTop_toBottomOf="@id/shazamzams_page_icon" />
 
     <View
-        android:id="@+id/horizontal_line"
+        android:id="@+id/right_horizontal_line_sign_up"
         android:layout_width="match_parent"
         android:layout_height="@dimen/line_width"
         android:background="@color/gunmetal"
