+++ app/src/main/res/layout/activity_my_profile.xml
@@ -46,7 +46,7 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <View
-        android:id="@+id/horizontal_line"
+        android:id="@+id/right_horizontal_line_sign_up"
         android:layout_width="match_parent"
         android:layout_height="@dimen/line_width"
         android:background="@color/gunmetal"
