+++ app/src/main/res/values/dimens.xml
@@ -26,4 +26,16 @@
     <dimen name="num_of_friends_rect_radius">20dp</dimen>
     <dimen name="shazamzams_textView_width">70dp</dimen>
     <dimen name="shazamzams_textView_height">22dp</dimen>
+    <dimen name="page_headline_width">120dp</dimen>
+    <dimen name="page_headline_height">70dp</dimen>
+    <dimen name="headline_top_margin">50dp</dimen>
+    <dimen name="line_height">2dp</dimen>
+    <dimen name="letsgo_button_width">110dp</dimen>
+    <dimen name="letsgo_button_height">55dp</dimen>
+    <dimen name="letsgo_bottom_margin">250dp</dimen>
+    <dimen name="sign_up_buttom_width">110dp</dimen>
+    <dimen name="sign_up_buttom_height">55dp</dimen>
+    <dimen name="decotation_height">250dp</dimen>
+    <dimen name="global_side_margin">20dp</dimen>
+    <dimen name="letsgo_bottom_margin_type2">200dp</dimen>
 </resources>
