+++ app/src/main/res/values/strings.xml
@@ -9,5 +9,8 @@
     <string name="num_of_listeners_text">5 Listeners</string>
     <string name="hum_length_text">TextView</string>
     <string name="hum_name_text">Song name</string>
+    <string name="email_hint">Email...</string>
+    <string name="password_hint">Password...</string>
+    <string name="user_name_hint">User Name...</string>
 
 </resources>
